libcpm: make the CP/M runtime link, and stop cpp forking on the target
Most of the translation layer was already here - bdos.s, the fcb
table, open/read/write/close/seek over BDOS calls - but it had never
been linked, so the parts that were missing or wrong had never had to
show themselves.  Four things were.

getuid.s was sitting in libc, which is the system-independent half.
It is a BDOS call - function 0x20, the user number - and libcpm's
open, read, seek and getargs all call it to switch user areas.  It
was in neither library's source list, so it was not built at all.
It moves to libcpm and gets built.

perror.c and sys_err.c were in the directory and not in CSRCS.
sys_err.c carries errno and the message table, which is what libu's
errno.s is on the other side.

fgetc.s ends "ld hl,EOF" and never had an equ for it; the assembler
took it for an external and the linker could not find one.  fseek.c
calls clreof(), which exists nowhere in this tree - libc's fseek
clears the flag directly, so this does too.  Neither had been
assembled or compiled before, which is how both survived.

seekraw is new.  On Micronix it is the bare kernel call and lseek is
built on it; here it is the other way round, because CP/M has no seek
and the read/write pointer lives in our own fcb, so it wraps the
lseek that maintains it.  cpp and pass2 both use it to move around
the .n name file.

perror now writes on the file descriptor instead of going through
putc and stderr.  Through stdio it dragged _iob and the buffering in,
which cpp deliberately avoids - it defines an empty _cleanup so exit()
does not pull that machinery in either - and libc's cleanup.c has
_cleanup in the same object as the _iob table, so the two collided.
libu's perror.s has always written the descriptor directly.

And cpp forked on the target.  The #ifdef CCC around -p chose which
exec to call but left fork() and wait() outside it, so the Z80 build
forked too.  Rendering the .x as text lives in xdump so its tables
stay out of cpp's image, and reaching it by forking is a host
convenience: -p is for reading what cpp did, it is not part of the
compile the driver runs, and a second process is not something to ask
of a target that has one program in memory.  It is host-only now, and
the target build says so rather than failing at a fork that cannot
work.  -E is untouched: it execs without forking, which is chaining,
which CP/M can do.

The driver grew what a .com actually needs - data and bss at the same
origin as text, so the image is contiguous, and the 16-byte object
header cut off the front - and make cpm builds all four passes into
cpm/bin.

Three of the four fit under a 2.2 BDOS at 0xE406: cpp ends 0xc7ca, c0
0xc75b, peep 0xd15f.  c1 ends 0xea2e and does not, which is the same
ceiling it is up against on Micronix.  None of this has been run:
there is no CP/M emulator here, so what is proven is that it compiles,
links and fits, and nothing beyond that.

+++ src/libcpm/sys_err.c
@@ -1,4 +1,4 @@
-short	errno;
+int	errno;
 char *	sys_err[] =
 {
 	"Unknown error",
@@ -9,7 +9,7 @@ char *	sys_err[] =
 	"Directory full",
 	"Seek past end of disk???",
 };
-short	sys_ner = sizeof sys_err / sizeof sys_err[0];
+int	sys_ner = sizeof sys_err / sizeof sys_err[0];
 /*
  * vim: set tabstop=4 shiftwidth=4 noexpandtab:
  */
